requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
boto3==1.37.30
botocore==1.37.30
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
frozenlist==1.5.0
future==1.0.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-documentai==3.3.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
line-bot-sdk==3.16.2
matplotlib-inline==0.1.7
motor==3.7.0
multidict==6.2.0
nest-asyncio==1.6.0
numpy==2.0.2
openai==1.70.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pyzmq==26.3.0
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
